Moved the components used for all pages to a separate directory. Renamed an alias
