+++ src/components/PortfolioContainer/index.js
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
-import NavTabs from "../NavTabs";
-import Header from "../pages/Header";
-import Footer from "../pages/Footer";
+import NavTabs from "../elements/NavTabs";
+import Header from "../elements/Header";
+import Footer from "../elements/Footer";
 import Portfolio from "../pages/Portfolio";
 import About from "../pages/About";
 import Resume from "../pages/Resume";
