+++ src/components/elements/Header/index.js
@@ -1,10 +1,10 @@
 import React from 'react'
-import { Header as HeaderWithImage, Image } from 'semantic-ui-react'
+import { Header as HeaderReact, Image } from 'semantic-ui-react'
 
 const Header = () => (
-  <HeaderWithImage as='h2'>
+  <HeaderReact as='h2'>
     <Image circular src='https://react.semantic-ui.com/images/avatar/large/patrick.png' /> Alexander Gorshkov
-  </HeaderWithImage>
+  </HeaderReact>
 )
 
 export default Header
